requirements.txt
# -*- coding: utf-8 -*-
tensorflow-federated
numpy
scipy
tqdm
matplotlib